+++ src/com/idbs/devassessment/solution/Term.java
@@ -96,13 +96,13 @@ public class Term {
 	 * @param xValue the given value of x
 	 * @return the evaluated term
 	 */
-	public int evaluate(int xValue) {
+	public long evaluate(int xValue) {
 		
 		if (this.multiplier == 0) { //anything multiplied by 0 is 0
 			return 0;
 		}
 		else {
-			int result = exponent(xValue, this.power); //calculate indices first (BIDMAS)
+			long result = exponent(xValue, this.power); //calculate indices first (BIDMAS)
 			if (result > this.multiplier) { //smaller value first so less loops required
 				result = multiply(result,this.multiplier); //multiply result by multiplier
 			}
@@ -125,8 +125,8 @@ public class Term {
 	 * @param a, b terms to be multiplied by each other
 	 * @return result of multiplication
 	 */
-	private int multiply(int a, int b) {
-		int result = a;
+	private long multiply(long a, long b) {
+		long result = a;
 		for (int i = 1; i < b; i++) { //b should be smaller than a to reduce amount of loops
 			result = result + a; //add a to itself b times
 		}
@@ -140,8 +140,8 @@ public class Term {
 	 * @param p the given value of power
 	 * @return result of x to the power of p
 	 */
-	private int exponent(int x, int p) {
-		int result = x; 
+	private long exponent(long x, int p) {
+		long result = x; 
 		
 		if(p == 0) { //anything to the power of 0 evaluates to 1
 			result = 1;
